Fixes in sent invite list tiles highlighting. Some refactoring to order inbox files
