+++ lib/widgets/screens/inbox/dismissible_tile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../constants/app_constants.dart';
+import '../../../constants/app_constants.dart';
 
 class DismissibleTile extends StatelessWidget {
   final String tileId;
