+++ lib/widgets/screens/inbox/inbox_invites_received.dart
@@ -6,7 +6,7 @@ import 'package:mm_flutter_app/providers/models/my_channel_invitations_model.dar
 import 'package:mm_flutter_app/utilities/router.dart';
 import 'package:mm_flutter_app/utilities/utility.dart';
 import 'package:mm_flutter_app/widgets/atoms/empty_state_message.dart';
-import 'package:mm_flutter_app/widgets/molecules/inbox_list_tile.dart';
+import 'package:mm_flutter_app/widgets/screens/inbox/inbox_list_tile.dart';
 import 'package:provider/provider.dart';
 
 import '../../../providers/models/scaffold_model.dart';
@@ -79,8 +79,9 @@ class _InboxInvitesReceivedScreenState extends State<InboxInvitesReceivedScreen>
       fullName: invitation.sender.fullName ?? '',
       date: invitation.createdAt.toLocal(),
       message: invitation.messageText ?? _l10n.inboxInvitesReceivedMessage,
-      highlightTile:
+      highlightTileTitle:
           true, // TODO - Highlight and show notification bubble only if unseen
+      highlightTileText: true,
       simplifyDate: true,
       onPressed: () => router.push(
         '${Routes.inboxInvitesReceived.path}/${invitation.id}',
